Refactor map out of Application into it's own component
