+++ src/components/Application/Application.js
@@ -1,8 +1,6 @@
 import React, {Component} from 'react';
-import { getApiKey } from '../../GoogleApiKey';
-
-import {Card, CardBody, CardHeader, Container, Button} from 'reactstrap';
-import GoogleMapReact from 'google-map-react';
+import Map from './Map';
+import {Card, CardBody, CardHeader, Button} from 'reactstrap';
 
 const DEBUG = true;
 
@@ -25,122 +23,18 @@ export default class Application extends Component {
     super(props);
 
     this.state = {
-      map: null,
-      oldBounds: {
-        ne_corner: null,
-        sw_corner: null
-      },
-      userActions: [],
-      center: {
-        lat: 40.559167,
-        lng: -105.078056
-      },
-      zoom: 11
+      userActions: []
     };
 
-    this.idleCallback = this.idleCallback.bind(this);
-    this.zoomChangedCallback = this.zoomChangedCallback.bind(this);
-    this.updateCenter = this.updateCenter.bind(this);
-    this.updateBounds = this.updateBounds.bind(this);
+
     this.addAction = this.addAction.bind(this);
-    this.logActions = this.logActions.bind(this);
-    this.handleApiLoaded = this.handleApiLoaded.bind(this);
+    this.getActionsAsString = this.getActionsAsString.bind(this);
   }
-
-  handleApiLoaded = (map, maps) => {
-    this.setState({map: map});
-    map.addListener('idle', this.idleCallback);
-    map.addListener('zoom_changed', this.zoomChangedCallback);
-  };
-
-  idleCallback() {
-    let currentCenter = this.state.map.getCenter();
-    let currentLat = currentCenter.lat();
-    let currentLng = currentCenter.lng();
-
-
-    if (currentLng < this.state.center.lng) {
-      if (DEBUG) {
-        console.log("PAN_LEFT");
-      }
-      this.addAction({ action: "PAN_LEFT", interval: 0.0 } );
-    }
-    else if (currentLng > this.state.center.lng) {
-      if (DEBUG) {
-        console.log("PAN_RIGHT");
-      }
-      this.addAction( { action: "PAN_RIGHT", interval: 0.0 } );
-    }
-    else if (currentLat > this.state.center.lat) {
-      if (DEBUG) {
-        console.log("PAN_UP");
-      }
-      this.addAction( { action: "PAN_UP", interval: 0.0 } );
-    }
-    else if (currentLat < this.state.center.lat) {
-      if (DEBUG) {
-        console.log("PAN_DOWN");
-      }
-      this.addAction( { action: "PAN_DOWN", interval: 0.0 } );
-    }
-
-    // Update old latitude and longitude
-    this.updateCenter(currentLat, currentLng);
-
-
-    if (this.state.oldBounds.ne_corner === null || this.state.oldBounds.sw_corner === null) {
-      let tempBounds = this.state.map.getBounds();
-      let tempSouthWestCorner = tempBounds.getSouthWest();
-      let tempNorthEastCorner = tempBounds.getNorthEast();
-
-      // Update oldBounds
-      this.updateBounds(tempNorthEastCorner, tempSouthWestCorner);
-    }
-  }
-
-  zoomChangedCallback() {
-    let tempBounds = this.state.map.getBounds();
-    let new_sw_corner = tempBounds.getSouthWest();
-    let new_ne_corner = tempBounds.getNorthEast();
-
-    if (this.state.oldBounds.ne_corner.lat < new_ne_corner.lat()) {
-      if (DEBUG) {
-        console.log("ZOOM_OUT");
-      }
-      this.addAction( { action: "ZOOM_OUT", interval: 0.0 });
-    }
-    else {
-      if (DEBUG) {
-        console.log("ZOOM_IN");
-      }
-      this.addAction( { action: "ZOOM_IN", interval: 0.0 });
-    }
-
-    // Update oldBounds
-    this.updateBounds(new_ne_corner, new_sw_corner);
-  }
-
 
   addAction(action) {
     let actions = this.state.userActions;
     actions.push(action);
     this.setState({'userActions': actions});
-  }
-
-  updateCenter(lat, lng) {
-    let oldCenter = this.state.center;
-    oldCenter.lat = lat;
-    oldCenter.lng = lng;
-    this.setState({center: oldCenter})
-  }
-
-  updateBounds(ne_corner, sw_corner) {
-    let oldBoundsCopy = this.state.oldBounds;
-
-    oldBoundsCopy.ne_corner = {lat: ne_corner.lat(), lng: ne_corner.lng()};
-    oldBoundsCopy.sw_corner = {lat: sw_corner.lat(), lng: sw_corner.lng()};
-
-    this.setState({'oldBounds': oldBoundsCopy});
   }
 
   getActionsAsString() {
@@ -155,6 +49,8 @@ export default class Application extends Component {
         }
       }
 
+      if (DEBUG) console.log(collectedActions);
+
       return collectedActions;
     }
   }
@@ -163,26 +59,13 @@ export default class Application extends Component {
   render() {
     return (
         <div className="application-width">
-          <Card>
-            <CardBody>
-              <div style={{ height: '80vh', width: '100%' }}>
-                <GoogleMapReact
-                    bootstrapURLKeys={{ key: getApiKey() }}
-                    defaultCenter={this.state.center}
-                    defaultZoom={this.state.zoom}
-                    yesIWantToUseGoogleMapApiInternals
-                    onGoogleApiLoaded={({ map, maps }) => this.handleApiLoaded(map, maps)}
-                >
-                </GoogleMapReact>
-              </div>
-            </CardBody>
-          </Card>
+          <Map addAction={this.addAction}/>
           <Card>
             <CardHeader>
               Control Panel
             </CardHeader>
             <CardBody>
-              <Button className='btn-csu' size="lg" onClick={this.logActions}>See Results</Button>
+              <Button className='btn-csu' size="lg" onClick={this.getActionsAsString}>See Results</Button>
             </CardBody>
           </Card>
         </div>
