+++ src/components/Margins/Header.js
@@ -5,22 +5,19 @@ import {HeaderLogoLarge, HeaderLogoMedium, HeaderLogoSmall} from './HeaderLogo';
 /* Renders a text heading above the application with useful information.
  */
 export default class Header extends Component{
-  constructor(props) {
-    super(props);
-  }
 
   render() {
     return(
         <div className="add-header-height">
           <div className="application-width">
             <div id="responsiveHeaderContainer">
-              <a href="http://colostate.edu" id="csuHeaderLink" target="_blank">
+              <a href="http://colostate.edu" id="csuHeaderLink" target="_blank" rel="noopener noreferrer">
                 <HeaderLogoLarge/>
                 <HeaderLogoMedium/>
                 <HeaderLogoSmall/>
               </a>
               <div id="responsiveLogoSubsystem">
-                <a href={"https://compsci.colostate.edu/"} id="csHeaderLink" target="_blank">
+                <a href={"https://compsci.colostate.edu/"} id="csHeaderLink" target="_blank" rel="noopener noreferrer">
                   <h1 className="larger-CSUtext-upper">
                     Computer Science
                   </h1>
